chore(quitte): cleanup unwarranted comments
The comments were written based on the assumption that observable collections would behave inconsistently for collections depending on identity-based equality (i.e. IdentityHashMap). However, this assumption is flawed since the contract of j.u.Collection mandates comparison by using equals. Thus, we might as well ignore minor inconsistencies for identity-based collections.

(Additionally, it is worth noting that fixing this is not a trivial task since there is no [sane] way to determine if a collection uses identity checks. We just assume that collections obey the spec. An "ObservableIdentityMap" might be a topic for another day.)